backend/requirements.txt
asgiref==3.8.1
build==1.2.2.post1
click==8.1.7
colorama==0.4.6
Django==4.2.17
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environ==1.0
gunicorn==23.0.0
importlib_metadata==8.5.0
packaging==24.1
pillow==11.0.0
pip-tools==7.4.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
pyproject_hooks==1.2.0
sqlparse==0.5.1
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.1
zipp==3.21.0